show directions to confectioner in external map
